+++ extern/json-c/arraylist.c
@@ -45,7 +45,7 @@ struct array_list *array_list_new2(array_list_free_fn *free_fn, int initial_size
 {
 	struct array_list *arr;
 
-	if (initial_size < 0 || (size_t)initial_size >= SIZE_T_MAX / sizeof(void *))
+	if (initial_size < 0 || (uint64)initial_size >= SIZE_T_MAX / sizeof(void *))
 		return NULL;
 	arr = (struct array_list *)malloc(sizeof(struct array_list));
 	if (!arr)
@@ -63,7 +63,7 @@ struct array_list *array_list_new2(array_list_free_fn *free_fn, int initial_size
 
 extern void array_list_free(struct array_list *arr)
 {
-	size_t i;
+	uint64 i;
 	for (i = 0; i < arr->length; i++)
 		if (arr->array[i])
 			arr->free_fn(arr->array[i]);
@@ -71,17 +71,17 @@ extern void array_list_free(struct array_list *arr)
 	free(arr);
 }
 
-void *array_list_get_idx(struct array_list *arr, size_t i)
+void *array_list_get_idx(struct array_list *arr, uint64 i)
 {
 	if (i >= arr->length)
 		return NULL;
 	return arr->array[i];
 }
 
-static int array_list_expand_internal(struct array_list *arr, size_t max)
+static int array_list_expand_internal(struct array_list *arr, uint64 max)
 {
 	void *t;
-	size_t new_size;
+	uint64 new_size;
 
 	if (max < arr->size)
 		return 0;
@@ -94,7 +94,7 @@ static int array_list_expand_internal(struct array_list *arr, size_t max)
 		if (new_size < max)
 			new_size = max;
 	}
-	if (new_size > (~((size_t)0)) / sizeof(void *))
+	if (new_size > (~((uint64)0)) / sizeof(void *))
 		return -1;
 	if (!(t = realloc(arr->array, new_size * sizeof(void *))))
 		return -1;
@@ -103,10 +103,10 @@ static int array_list_expand_internal(struct array_list *arr, size_t max)
 	return 0;
 }
 
-int array_list_shrink(struct array_list *arr, size_t empty_slots)
+int array_list_shrink(struct array_list *arr, uint64 empty_slots)
 {
 	void *t;
-	size_t new_size;
+	uint64 new_size;
 
 	if (empty_slots >= SIZE_T_MAX / sizeof(void *) - arr->length)
 		return -1;
@@ -125,9 +125,9 @@ int array_list_shrink(struct array_list *arr, size_t empty_slots)
 	return 0;
 }
 
-int array_list_insert_idx(struct array_list *arr, size_t idx, void *data)
+int array_list_insert_idx(struct array_list *arr, uint64 idx, void *data)
 {
-	size_t move_amount;
+	uint64 move_amount;
 
 	if (idx >= arr->length)
 		return array_list_put_idx(arr, idx, data);
@@ -147,7 +147,7 @@ int array_list_insert_idx(struct array_list *arr, size_t idx, void *data)
 }
 
 //static inline int _array_list_put_idx(struct array_list *arr, size_t idx, void *data)
-int array_list_put_idx(struct array_list *arr, size_t idx, void *data)
+int array_list_put_idx(struct array_list *arr, uint64 idx, void *data)
 {
 	if (idx > SIZE_T_MAX - 1)
 		return -1;
@@ -177,7 +177,7 @@ int array_list_add(struct array_list *arr, void *data)
 	/* Repeat some of array_list_put_idx() so we can skip several
 	   checks that we know are unnecessary when appending at the end
 	 */
-	size_t idx = arr->length;
+	uint64 idx = arr->length;
 	if (idx > SIZE_T_MAX - 1)
 		return -1;
 	if (array_list_expand_internal(arr, idx + 1))
@@ -198,14 +198,14 @@ void *array_list_bsearch(const void **key, struct array_list *arr,
 	return bsearch(key, arr->array, arr->length, sizeof(arr->array[0]), compar);
 }
 
-size_t array_list_length(struct array_list *arr)
+uint64 array_list_length(struct array_list *arr)
 {
 	return arr->length;
 }
 
-int array_list_del_idx(struct array_list *arr, size_t idx, size_t count)
+int array_list_del_idx(struct array_list *arr, uint64 idx, uint64 count)
 {
-	size_t i, stop;
+	uint64 i, stop;
 
 	/* Avoid overflow in calculation with large indices. */
 	if (idx > SIZE_T_MAX - count)
